add tests for moscot kernel
