+++ cellrank/kernels/_moscot_kernel.py
@@ -8,6 +8,9 @@ import scanpy as sc
 import anndata
 from cellrank.kernels._transport_map_kernel import TransportMapKernel
 
+import numpy as np
+import scipy.sparse as sp
+
 __all__ = ["MoscotKernel"]
 
 
@@ -30,9 +33,9 @@ class MoscotKernel(TransportMapKernel):
         """Load from a moscot model."""
         return MoscotKernel(
             adata=problem.adata,
+            problem=problem,
             time_key=problem.temporal_key,
             backward=backward,
-            problem=problem,
             **kwargs,
         )
 
@@ -46,16 +49,21 @@ class MoscotKernel(TransportMapKernel):
         ] = "connectivities",
         conn_weight: Optional[float] = None,
         conn_kwargs: Mapping[str, Any] = MappingProxyType({}),
+        **kwargs: Any,
     ) -> "MoscotKernel":
         """Write transport matrices to kernel."""
+        cache_params = dict(kwargs)
+        cache_params["threshold"] = threshold
+        cache_params["self_transitions"] = self_transitions
+        if self._reuse_cache(cache_params):
+            return self
         if tmaps is None:
             if self.problem is None:
                 raise ValueError(
                     "If `MoscotKernel.problem` is `None`, `tmaps` must be provided."
                 )
-            self._moscot_to_cr_tmaps({p.solution for p in self.problem.problems})
-        else:
-            self._tmaps = self._moscot_to_cr_tmaps(tmaps)
+            tmaps = {p: self.problem[p].solution for p in self.problem.problems}
+        self._tmaps = self._moscot_to_cr_tmaps(tmaps)
 
         if threshold is not None:
             self._threshold_transport_maps(
@@ -83,10 +91,13 @@ class MoscotKernel(TransportMapKernel):
             src_mask = self.adata.obs[self._time_key] == t1
             tgt_mask = self.adata.obs[self._time_key] == t2
             tmats[t1, t2] = anndata.AnnData(
-                output.transport_matrix,
+                output.transport_matrix
+                if isinstance(output.transport_matrix, sp.csr_matrix)
+                else np.asarray(output.transport_matrix),
                 obs=self.adata[src_mask].obs,
                 var=self.adata[tgt_mask].obs,
             )
+        return tmats
 
     @property
     def problem(self) -> Type[CompoundProblem]:
