+++ requirements.txt
@@ -1,5 +1,6 @@
 anndata>=0.8,<0.9
 docrep>=0.3.0
+git+https://github.com/theislab/moscot
 joblib>=0.13.1
 matplotlib>=3.3.0
 networkx>=2.2
@@ -15,4 +16,3 @@ scvelo>=0.2.4
 seaborn>=0.10.0
 setuptools>=41.0.1
 wrapt>=1.12.1
-git+https://github.com/theislab/moscot
