+++ tests/test_external.py
@@ -1,4 +1,4 @@
-from typing import Optional, Union
+from typing import Union, Optional
 
 import pytest
 
@@ -48,9 +48,6 @@ def _statot_not_installed() -> bool:
 
 wot_not_installed_skip = pytest.mark.skipif(
     _wot_not_installed(), reason="WOT is not installed."
-)
-moscot_not_installed_skip = pytest.mark.skipif(
-    _moscot_not_installed(), reason="moscot is not installed."
 )
 statot_not_installed_skip = pytest.mark.skipif(
     _statot_not_installed(), reason="statOT is not installed."
@@ -385,155 +382,3 @@ class TestGetMarkers:
                 np.testing.assert_array_equal(
                     markers, [marker.upper() for marker in markers]
                 )
-
-
-@moscot_not_installed_skip
-class TestMoscotKernel:
-    from moscot.problems import LineageProblem, TemporalProblem, SpatioTemporalProblem
-
-    def test_init_from_adata(self, adata_moscot: AnnData):
-        pass
-
-    @pytest.mark.parametrize("backward", [True, False])
-    def test_init_from_TemporalProblem(
-        self, moscot_tp: TemporalProblem, backward: bool
-    ):
-        from moscot.problems import TemporalProblem
-
-        mk = MoscotKernel.load(moscot_tp, backward=backward)
-        assert mk.transport_maps is None
-        assert mk.obs is None
-        assert isinstance(mk.problem, TemporalProblem)
-
-    @pytest.mark.parametrize("backward", [True, False])
-    def test_init_from_SpatioTemporalProblem(
-        self, moscot_stp: SpatioTemporalProblem, backward: bool
-    ):
-        from moscot.problems import SpatioTemporalProblem
-
-        mk = MoscotKernel.load(moscot_stp, backward=backward)
-        assert mk.transport_maps is None
-        assert mk.obs is None
-        assert isinstance(mk.problem, SpatioTemporalProblem)
-
-    @pytest.mark.parametrize("backward", [True, False])
-    def test_init_from_LineageProblem(self, moscot_lp: LineageProblem, backward: bool):
-        from moscot.problems import LineageProblem
-
-        mk = MoscotKernel.load(moscot_lp, backward=backward)
-        assert mk.transport_maps is None
-        assert mk.obs is None
-        assert isinstance(mk.problem, LineageProblem)
-
-    @pytest.mark.parametrize("backward", [True, False])
-    def test_compute_transition_matrix(
-        self, moscot_tp: TemporalProblem, backward: bool
-    ):
-        mk = MoscotKernel.load(moscot_tp, backward=backward)
-        assert mk.transport_maps is None
-        assert mk.obs is None
-
-        mk = mk.compute_transition_matrix()
-        assert isinstance(mk.transport_maps, dict)
-        assert len(mk.transport_maps) == 2
-        assert mk.transition_matrix is not None
-
-    @pytest.mark.parametrize("backward", [True, False])
-    def test_compute_transition_matrix_from_output(
-        self, moscot_tp: TemporalProblem, backward: bool
-    ):
-        mk = MoscotKernel.load(moscot_tp, backward=backward)
-        assert mk.transport_maps is None
-        assert mk.obs is None
-
-        tmaps = {p.solution for p in moscot_tp.problems}
-        mk = mk.compute_transition_matrix(tmaps=tmaps)
-        assert isinstance(mk.transport_maps, dict)
-        assert len(mk.transport_maps) == 2
-        assert mk.transition_matrix is not None
-
-    def test_invalid_solver_kwargs(self, adata_large: AnnData):
-        mk = MoscotKernel(adata_large, time_key="age(days)")
-        with pytest.raises(TypeError, match="unexpected keyword argument 'foo'"):
-            mk.compute_transition_matrix(foo="bar")
-
-    def test_inversion_updates_adata(self, adata_large: AnnData):
-        key = "age(days)"
-        mk = MoscotKernel(adata_large, time_key=key)
-        assert is_categorical_dtype(adata_large.obs[key])
-        assert adata_large.obs[key].cat.ordered
-        np.testing.assert_array_equal(mk.experimental_time, adata_large.obs[key])
-        orig_time = mk.experimental_time
-
-        mk = ~mk
-
-        assert mk.transition_matrix is None
-        assert mk.transport_maps is None
-
-        inverted_time = mk.experimental_time
-        assert is_categorical_dtype(adata_large.obs[key])
-        assert adata_large.obs[key].cat.ordered
-        np.testing.assert_array_equal(mk.experimental_time, adata_large.obs[key])
-        np.testing.assert_array_equal(
-            orig_time.cat.categories, inverted_time.cat.categories
-        )
-        np.testing.assert_array_equal(orig_time.index, inverted_time.index)
-
-        with pytest.raises(AssertionError):
-            np.testing.assert_array_equal(orig_time, inverted_time)
-
-    @pytest.mark.parametrize("cmap", ["inferno", "viridis"])
-    def test_update_colors(self, adata_large: AnnData, cmap: str):
-        ckey = "age(days)_colors"
-        _ = MoscotKernel(adata_large, time_key="age(days)", cmap=cmap)
-
-        colors = adata_large.uns[ckey]
-        cmap = get_cmap(cmap)
-
-        assert isinstance(colors, np.ndarray)
-        assert colors.shape == (2,)
-        np.testing.assert_array_equal(colors, [to_hex(cmap(0)), to_hex(cmap(cmap.N))])
-
-    @pytest.mark.parametrize("st", list(SelfTransitions))
-    def test_self_transitions(self, moscot_tp: TemporalProblem, st: SelfTransitions):
-        et, lt = 12.0, 35.0
-        key = "age(days)"
-        conn_kwargs = {"n_neighbors": 11}
-
-        mk = MoscotKernel.load(moscot_tp, time_key=key).compute_transition_matrix(
-            self_transitions=st,
-            conn_weight=0.2,
-            conn_kwargs=conn_kwargs,
-            threshold=None,
-        )
-
-        ixs = np.where(moscot_tp.adata.obs[key] == lt)[0]
-        T = mk.transition_matrix[ixs, :][:, ixs].A
-
-        assert mk.params["self_transitions"] == str(st)
-        if st == SelfTransitions.UNIFORM:
-            np.testing.assert_allclose(T, np.ones_like(T) / float(len(ixs)))
-        elif st == SelfTransitions.DIAGONAL:
-            np.testing.assert_allclose(T, np.eye(len(ixs)))
-        elif st in (SelfTransitions.CONNECTIVITIES, SelfTransitions.ALL):
-            adata_subset = moscot_tp.adata[moscot_tp.adata.obs[key] == lt]
-            sc.pp.neighbors(adata_subset, **conn_kwargs)
-            T_actual = (
-                ConnectivityKernel(adata_subset)
-                .compute_transition_matrix()
-                .transition_matrix.A
-            )
-            np.testing.assert_allclose(T, T_actual)
-            if st == SelfTransitions.ALL:
-                ixs = np.where(moscot_tp.adata.obs[key] == et)[0]
-                T = mk.transition_matrix[ixs, :][:, ixs].A
-                adata_subset = moscot_tp[moscot_tp.adata.obs[key] == et]
-                sc.pp.neighbors(adata_subset, **conn_kwargs)
-                T_actual = (
-                    ConnectivityKernel(adata_subset)
-                    .compute_transition_matrix()
-                    .transition_matrix.A
-                )
-                np.testing.assert_allclose(T, 0.2 * T_actual)
-        else:
-            raise NotImplementedError(st)
